bot/requirements.txt
beautifulsoup4~=4.9.0
requests~=2.22.0
python-dotenv==0.18.0
python-telegram-bot==13.6